implement clock disparity for attestation validation (#1568)
This implements disparity, resolving a part of
https://github.com/status-im/nim-beacon-chain/issues/1367

* make BeaconTime a duration for fractional seconds
* factor out attestation/aggregate validation
* simplify recording of queued attestations
* simplify attestation signature check
* fix blocks_received metric
* add some trivial validation tests
* remove unresolved attestation table - attestations for unknown blocks
are dropped instead (cannot verify their signature)

+++ beacon_chain/beacon_node.nim
@@ -867,8 +867,8 @@ proc start(node: BeaconNode) =
     version = fullVersionStr,
     nim = shortNimBanner(),
     timeSinceFinalization =
-      int64(finalizedHead.slot.toBeaconTime()) -
-      int64(node.beaconClock.now()),
+      finalizedHead.slot.toBeaconTime() -
+      node.beaconClock.now(),
     head = shortLog(head),
     finalizedHead = shortLog(finalizedHead),
     SLOTS_PER_EPOCH,
